+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/CaseEventReceiver.java
@@ -1,26 +1,17 @@
 package uk.gov.ons.ctp.integration.rhsvc.message;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.integration.annotation.MessageEndpoint;
-import org.springframework.integration.annotation.ServiceActivator;
+import uk.gov.ons.ctp.integration.rhsvc.domain.CaseEvent;
 
 /**
- * Service implementation responsible for receipt of Case Events. See Spring Integration flow for
- * details of in bound queue.
+ * Interface for the receipt of Case Events. See Spring Integration flow for details of in bound
+ * queue.
  */
-@MessageEndpoint
-public class CaseEventReceiver {
-
-  @Autowired private RespondentEventPublisher publisher;
+public interface CaseEventReceiver {
 
   /**
-   * Message end point for events from Response Management. At present sends straight to publisher
-   * to prove messaging setup.
+   * Method called with the deserialised message received from the Case service
    *
-   * @param event CaseEvent message from Response Management
+   * @param event CasesEvent object received
    */
-  @ServiceActivator(inputChannel = "acceptCaseEvent")
-  public void acceptCaseEvent(CaseEvent event) {
-    publisher.sendEvent(event);
-  }
+  void acceptCaseEvent(CaseEvent event);
 }
