+++ src/main/resources/springintegration/respondent-service-event-outbound.xml
@@ -7,8 +7,7 @@
   http://www.springframework.org/schema/rabbit
   http://www.springframework.org/schema/rabbit/spring-rabbit-1.6.xsd">
   
-    <bean id="respondentEventJsonMessageConverter" class="org.springframework.amqp.support.converter.Jackson2JsonMessageConverter">
-    </bean>
+    <bean id="respondentEventJsonMessageConverter" class="org.springframework.amqp.support.converter.Jackson2JsonMessageConverter"/>
 
     <rabbit:template id="respondentEventRabbitTemplate" connection-factory="connectionFactory" exchange="respondent-outbound-exchange" routing-key="Respondent.Event.binding"
                      message-converter="respondentEventJsonMessageConverter" channel-transacted="true"/>
